Update README.md
The Stripe Payment Terminal is an Odoo module that allows you to accept card payments using a Stripe terminal. It allows you to process payments quickly and securely, and is especially useful for businesses that need to process a high volume of transactions in a short amount of time. To use the Stripe Payment Terminal in app, you will need to have a Stripe account and a compatible Stripe terminal. Once you have these set up, you can configure the Stripe Payment Terminal module in appto start accepting payments.